rename service package to 'services'
